Upload image to server VDO 08

+++ resources/js/Pages/Store.vue
@@ -5,8 +5,15 @@
 
     <div v-if="ShowForm">
         {{ FormStore }}
+        <hr>
         <div class="row">
-            <div class="col-md-4">Image</div>
+            <div class="col-md-4 text-center" style=" position:relative">
+                <button type="button" v-if="FormStore.image" @click="RemoveImg()" class="btn rounded-pill btn-icon btn-danger btimg">
+                    <i class='bx bx-x-circle fs-3'></i>
+              </button>
+                <img :src="ImagePreview" @click="$refs.img_store.click()" style=" width:70%" class=" cursor-pointer">
+                <input type="file" ref="img_store" style=" display:none" @change="onSelect($event)" >
+            </div>
             <div class="col-md-8">
                 <div class="mb-2">
                     <label  class="form-label">ຊື່ສິນຄ້າ:</label>
@@ -14,13 +21,13 @@
                 </div>
                 <div class="mb-2">
                     <label  class="form-label">ຈຳນວນ:</label>
-                    <input type="text" class="form-control" v-model="FormStore.qty"  placeholder="..." >
+                    <cleave :options="options"  class="form-control" v-model="FormStore.qty"  placeholder="..." />
                 </div>
                 <div class="row">
                     <div class=" col-md-6 mb-2">
                         <label  class="form-label">ລາຄາຊື້:</label>
                         <div class="input-group">
-                            <input type="text" class="form-control" v-model="FormStore.price_buy" placeholder="..." >
+                            <cleave :options="options" class="form-control" v-model="FormStore.price_buy" placeholder="..." />
                             <span class="input-group-text" id="basic-addon11">ກີບ</span>
                         </div>
 
@@ -28,7 +35,7 @@
                     <div class="col-md-6 mb-2">
                         <label  class="form-label">ລາຄາຂາຍ:</label>
                         <div class="input-group">
-                            <input type="text" class="form-control" v-model="FormStore.price_sell"  placeholder="..." >
+                            <cleave :options="options" class="form-control" v-model="FormStore.price_sell"  placeholder="..." />
                             <span class="input-group-text" id="basic-addon11">ກີບ</span>
                         </div>
                     </div>
@@ -44,21 +51,22 @@
     <div class="table-responsive text-nowrap" v-else>
         <div class=" d-flex justify-content-between mb-2">
             <div class=" d-flex">
-                <div class=" d-flex align-items-center me-2 ">
-                    <i class='bx bx-sort-down fs-3'></i>
+                <div class=" d-flex align-items-center me-2 cursor-pointer " @click="ChangeSort()" >
+                    <i class='bx bx-sort-down fs-3' v-if="Sort=='desc'"></i>
+                    <i class='bx bx-sort-up fs-3' v-if="Sort=='asc'"></i>
                 </div>
                 
-                <select class=" form-select">
-                    <option>5</option>
-                    <option>10</option>
-                    <option>20</option>
-                    <option>30</option>
+                <select v-model="PerPage" class=" form-select" @change="GetStore()" >
+                    <option value="5" >5</option>
+                    <option value="10" >10</option>
+                    <option value="20" >20</option>
+                    <option value="30" >30</option>
                 </select>
             </div>
             <div class=" d-flex">
                 <div class="input-group me-2">
-                    <input type="text" class="form-control" placeholder="ຄົ້ນຫາ...">
-                    <button class="btn btn-primary p-2 " type="button" ><i class='bx bx-search-alt fs-5'></i></button>
+                    <input type="text" class="form-control" v-model="Search" @keyup.enter="GetStore()"  placeholder="ຄົ້ນຫາ...">
+                    <button class="btn btn-primary p-2 " type="button" @click="GetStore()" ><i class='bx bx-search-alt fs-5'></i></button>
                 </div>
                 <button type="button" class="btn btn-info" @click="AddStore()" >ເພີ່ມຂໍ້ມູນ</button>
             </div>
@@ -67,12 +75,12 @@
       <table class="table table-bordered">
         <thead>
           <tr>
-            <th>ID</th>
-            <th>ຮູບພາບ</th>
-            <th>ຊື່ສິນຄ້າ</th>
-            <th>ຈຳນວນ</th>
-            <th>ລາຄາຊື້</th>
-            <th>ຈັດການ</th>
+            <th class="text-center">ID</th>
+            <th class="text-center">ຮູບພາບ</th>
+            <th class="text-center">ຊື່ສິນຄ້າ</th>
+            <th class="text-center">ຈຳນວນ</th>
+            <th class="text-center">ລາຄາຊື້</th>
+            <th class="text-center">ຈັດການ</th>
           </tr>
         </thead>
         <tbody>
@@ -83,8 +91,8 @@
             <td>
              {{ item.qty }}
             </td>
-            <td>{{ item.price_buy }}</td>
-            <td>
+            <td class="text-end">{{ formatPrice(item.price_buy) }}</td>
+            <td  class="text-center">
               <div class="dropdown">
                 <button type="button" class="btn p-0 dropdown-toggle hide-arrow" data-bs-toggle="dropdown"><i class="bx bx-dots-vertical-rounded"></i></button>
                 <div class="dropdown-menu">
@@ -98,7 +106,7 @@
           
         </tbody>
       </table>
-
+      <button @click="showAlert">Hello world</button>
       <Pagination :pagination="StoreData" :offset="4" @paginate="GetStore($event)" />
     </div>
 
@@ -119,6 +127,7 @@ export default {
     },
     data() {
         return {
+            ImagePreview: window.location.origin + '/assets/img/upload.jpg',
             ShowForm: false,
             FormType: true,
             StoreData:[],
@@ -129,7 +138,20 @@ export default {
                 qty:'',
                 price_buy:'',
                 price_sell:''
-            }
+            },
+            Sort:'desc',
+            PerPage: 5,
+            Search:'',
+            options: {
+                  numeral: true,
+                  numeralPositiveOnly: true,
+                  noImmediatePrefix: true,
+                  rawValueTrimPrefix: true,
+                  numeralIntegerScale: 10,
+                  numeralDecimalScale: 2,
+                  numeralDecimalMark: '.',
+                  delimiter: ','
+                }
         }
     },
     computed:{
@@ -142,9 +164,46 @@ export default {
         }
     },
     methods:{
+        RemoveImg(){
+            this.FormStore.image = ''
+            this.ImagePreview = window.location.origin + '/assets/img/upload.jpg'
+        },
+        showAlert() {
+      // Use sweetalert2
+      this.$swal({
+                        // title: "ການລຶບຂໍ້ມູນບໍ່ສຳເລັດ!",
+                        text: "Your file has been deleted.",
+                        showConfirmButton:false,
+                        icon: "success",
+                        timer: 2500
+                    });
+            },
+        onSelect(event){
+            // console.log(event)
+
+            this.FormStore.image = event.target.files[0];
+            let reader = new FileReader();
+            reader.readAsDataURL(this.FormStore.image);
+            reader.addEventListener("load", function(){
+                this.ImagePreview = reader.result
+            }.bind(this,false));
+        },
+        formatPrice(value) {
+            let val = (value / 1).toFixed(0).replace(",", ".");
+            return val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
+        },
+        ChangeSort(){
+            if(this.Sort == 'asc'){
+                this.Sort = 'desc'
+            } else {
+                this.Sort = 'asc'
+            }
+            this.GetStore()
+        },
         AddStore(){
             this.ShowForm = true
             this.FormType = true
+            this.ImagePreview = window.location.origin + '/assets/img/upload.jpg'
         },
         CancelStore(){
             this.ShowForm = false
@@ -165,23 +224,64 @@ export default {
         },
         DelStore(id){
 
-            axios.delete(`api/store/delete/${id}`,{ headers:{ Authorization: 'Bearer ' + this.store.get_token } }).then((res)=>{
+
+            this.$swal({
+            title: "ທ່ານແນ່ໃຈບໍ່?",
+            text: "ທີ່ຈະລຶບຂໍ້ມູນນີ້!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "ຍືນຍັນລຶບ!",
+            cancelButtonText:"ຍົກເລີກ"
+            }).then((result) => {
+            if (result.isConfirmed) {
+
+
+                axios.delete(`api/store/delete/${id}`,{ headers:{ Authorization: 'Bearer ' + this.store.get_token } }).then((res)=>{
                 if(res.data.success){
+
+                    this.$swal({
+                        title: "Deleted!",
+                        text: "Your file has been deleted.",
+                        icon: "success"
+                    });
+
                     // ອັບເດດລາຍການ
                     this.GetStore()
-                } else {
+                    } else {
 
-                }
-            }).catch((error)=>{
-                console.log(error)
-            })
+                        this.$swal({
+                        title: "ການລຶບຂໍ້ມູນບໍ່ສຳເລັດ!",
+                        text: res.data.message,
+                        icon: "error"
+                    });
+
+                    }
+                }).catch((error)=>{
+                    console.log(error)
+                    if( typeof error.response !=='undefined'){
+                        this.$swal({
+                            title: "ການລຶບຂໍ້ມູນບໍ່ສຳເລັດ!",
+                            text: error.response.data.message,
+                            icon: "error"
+                        });
+                    }
+                })
+                
+                
+            }
+            });
+
+
+           
         },
         SaveStore(){
 
             if(this.FormType){
                 // ເພີ່ມຂໍ້ມູນໃໝ່
 
-                axios.post('api/store/add',this.FormStore,{ headers:{ Authorization: 'Bearer ' + this.store.get_token } }).then((res)=>{
+                axios.post('api/store/add',this.FormStore,{ headers:{ "Content-Type":"multipart/form-data", Authorization: 'Bearer ' + this.store.get_token } }).then((res)=>{
 
                     // console.log(res)
 
@@ -199,18 +299,37 @@ export default {
                             // ອັບເດດລາຍການ
                             this.GetStore()
 
+                            this.$swal({
+                                text: res.data.message,
+                                showConfirmButton:false,
+                                icon: "success",
+                                timer: 2500
+                            });
+
                         } else {
-                            
+                            this.$swal({
+                                text: res.data.message,
+                                showConfirmButton:false,
+                                icon: "error",
+                                timer: 2500
+                            });
                         }
 
                     }).catch((error)=>{
                     console.log(error)
+                        if( typeof error.response !=='undefined'){
+                            this.$swal({
+                                title: "ມີຂໍ້ຜິດຜາດ!",
+                                text: error.response.data.message,
+                                icon: "error"
+                            });
+                        }
                     })
 
             } else {
 
                 // ອັບເດດຂໍ້ມູນ
-                axios.post(`api/store/update/${this.EditID}`, this.FormStore,{ headers:{ Authorization: 'Bearer ' + this.store.get_token } }).then((res)=>{
+                axios.post(`api/store/update/${this.EditID}`, this.FormStore,{ headers:{ "Content-Type":"multipart/form-data", Authorization: 'Bearer ' + this.store.get_token } }).then((res)=>{
                   
                     if(res.data.success){
 
@@ -226,12 +345,34 @@ export default {
                             // ອັບເດດລາຍການ
                             this.GetStore()
 
+                            this.$swal({
+                                text: res.data.message,
+                                showConfirmButton:false,
+                                icon: "success",
+                                timer: 2500
+                            });
+
                     } else {
+
+
+                        this.$swal({
+                                text: res.data.message,
+                                showConfirmButton:false,
+                                icon: "error",
+                                timer: 2500
+                            });
 
                     }
 
                 }).catch((error)=>{
                     console.log(error);
+                    if( typeof error.response !=='undefined'){
+                            this.$swal({
+                                title: "ມີຂໍ້ຜິດຜາດ!",
+                                text: error.response.data.message,
+                                icon: "error"
+                            });
+                        }
                 })
 
             }
@@ -240,21 +381,39 @@ export default {
         },
         GetStore(page){
 
-            axios.get(`api/store?page=${page}`,{ headers:{ Authorization: 'Bearer ' + this.store.get_token } }).then((res)=>{
+            axios.get(`api/store?page=${page}&sort=${this.Sort}&perpage=${this.PerPage}&search=${this.Search}`,{ headers:{ Authorization: 'Bearer ' + this.store.get_token } }).then((res)=>{
 
                 this.StoreData = res.data;
 
             }).catch((error)=>{
                 console.log(error)
+                if( typeof error.response !=='undefined'){
+                            this.$swal({
+                                title: "ມີຂໍ້ຜິດຜາດ!",
+                                text: error.response.data.message,
+                                icon: "error"
+                            });
+                        }
             })
 
         }
     },
     created(){
         this.GetStore()
+    },
+    watch:{
+        Search(){
+            if(this.Search == ''){
+                this.GetStore()
+            }
+        }
     }
 }
 </script>
-<style lang="">
-    
+<style >
+    .btimg{
+        position: absolute;
+        top: 0px;
+        right: 10px;
+    }
 </style>
